Cambion y avances hasta Registro de pedidos en borrador

+++ backend/commons-microservicios/src/main/java/com/pymes/commons/services/CommonService.java
@@ -6,10 +6,10 @@ public interface CommonService<E> {
 
     public Iterable<E> findAll();
 
-    public Optional<E> findById(Long id);
+    public Optional<E> findById(Integer id);
 
     public E save(E entity);
 
-    public void deleteById(Long id);
+    public void deleteById(Integer id);
 
 }
