+++ backend/microservicios-proveedores/src/main/java/com/pymes/compras/app/proveedores/models/repository/ProveedorRepository.java
@@ -2,8 +2,8 @@ package com.pymes.compras.app.proveedores.models.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.pymes.commons.proveedores.models.entity.Proveedor;
+import com.pymes.commons.models.entity.Proveedor;
 
-public interface ProveedorRepository extends CrudRepository<Proveedor, Long>{
+public interface ProveedorRepository extends CrudRepository<Proveedor, Integer>{
     
 }
